:sparkles: [#5] Extract scaffolding of landing page
This is a very pragmatic approach with some explicit assumptions about
base templates and blocks being present, so it should be coordinated
with the (base) templates in default project.

For now the master.html template is not included at all, as this is
something that often requires some tweaks in each project and we want
to keep that flexibility.