:sparkles: [#5] Port the landing page/components CSS
Decided not to use a CSS pre-processor and instead use native CSS
nesting. The target audience for APIs are developers/technical people
of which we can reasonably assume that they can install a browser
newer than January 2024.

The CSS itself is not that exciting that we expect large benefits from
minifications, and it keeps our toolchain lean.