:sparkles: [#5] Port over tabs Javascript
I'm sure there must also exist a pure CSS solution for this, something
to explore for the future...

For the time being, the minimal JS is included in the library and
automatically loaded in the base template. Most of the API projects
can now probably drop their (public facing) JS build toolchain.