:technologist: Improve DX when running tests with pytest directly
If you have all the extras installed locally for
local development, running pytest would fail on
the errors because it assumes it's running in
the 'base' test environment in tox.

Now, the tests are skipped in those situations,
which adds a little bit more risk in being
careful which extras you setup in tox.ini.