:white_check_mark: [#5] Add testapp API landing page
Added a simple template inheritance and other
necessary bits to make the testapp look similar
enough to a default-project project.

The basic CSS is loaded and the tabs are interactive,
this is now a good starting point to write some
accessible playwright tests and improve the JS
and markup accordingly.