:wheelchair: [#5] Implement tabs on landing page in an accessible way
All the credit goes to the MDN docs:
https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Reference/Roles/tab_role

The markup, JS and CSS are updated so that it's accessible and
understood by screenreaders. Following Chris' feedback, the JS is
updated to use accessible queries to select elements, which helps lock
in the a11y requirements.

The CSS is modified slightly to account for the (new) button element
which comes with some default background styling out of the box,
and the 'active' modifier is no longer required on the tab panels.

Note that each tab element is now a button because that handles the
Enter and spacebar keys properly to activate the associated tab panel.