:white_check_mark: [#6] Add tests for remaining edge cases
It's possible a project opts in to a different kind of storage, e.g.
to get files on S3 and in those situations, a storage that's not a
subclass of FileSystemStorage is typically used. We can simulate the
expected behaviour by using the in memory storage shipped by Django.

Note that this storage does not set up a location, which the default
static files storage does (making sure every asset URL is prefixed
with settings.STATIC_URL). As such, the {% static %} template tag
effectively does nothing with this in-memory storage.