:bricks: [open-formulieren/open-forms#5451] Adapt initialization code for uwsgi
The exporter and provider initialization must happen in the post fork
hook of uwsgi, otherwise the threading.Lock instances are shared and
lead to a deadlock because uwsgi is a pre-forking server.

The same code is called for initialization, so we can rely on the
uwsgi module injection when the code runs in uwsgi to wire it up
correctly to the postfork (when the worker thread is properly set
up).

This did (after many hours of debugging) reveal that the django
instrumentation has to be performed *outside* of the postfork hook,
otherwise the metrics from the middleware don't get properly
initialized and no metrics are collected (and in turn, exported) at
all.

The double-initialization-prevention based on module globals also does
not work in this model, especially if the wsgi module gets loaded in
the master process and then the memory/state gets copied while forking,
which would incorrectly mark the otel machinery as initialized in the
forked process, while it's not, leading to the background thread not
starting at all.

Some more interesting findings while debugging this:

* --lazy-apps (ChatGPT was convinced this was needed) seems to break the
  postfork decorator entirely
* --py-call-uwsgi-fork-hooks is hinted at by the sentry-sdk, but it
  leads to segfaults when --lazy-apps is not enabled (at least in the
  case of Open Forms)
* we can use some more sane defaults in our uwsgi setup in default
  project
* gunicorn would likely *not* make this easier