:sparkles: [#7] Port the config helper
Ported the config helper function to read configuration from
the environment, but now with extensive typing support.

This way the intention of the resulted types can be declared
in settings files, and have these validated. It also
covers the possible invocations/combinations of (keyword)
arguments.

Finally, the default argument must be explicitly provided
as a kwarg, and a positional argument is no longer
permitted. This is to improve readability and avoid
potential confusion.